feat(merge-requests): implement mr-04 through mr-16
Push enforcement, storage merge ops, MR domain with approvals and
server-side merge, collaboration UI, settings, notifications, and e2e script.

+++ applications/OpenGitBase.Api/Controllers/RepositoryProtectedBranchRulesController.cs
@@ -201,7 +201,7 @@ public class RepositoryProtectedBranchRulesController : ControllerBase
             AllowedPushUserIds = request.AllowedPushUserIds.Select(UserId.From).ToList(),
             RequireMergeRequest = request.RequireMergeRequest,
             RequiredApprovalCount = request.RequiredApprovalCount,
-            MergeRoleThreshold = (int)request.MergeRoleThreshold,
+            MergeRoleThreshold = request.MergeRoleThreshold,
             ForcePushPolicy = request.ForcePushPolicy,
             DismissApprovalsOnPush = request.DismissApprovalsOnPush,
             LockedMergeStrategy = request.LockedMergeStrategy,
